ДЗ_12 рефакторинг. Переделал на дженерики.

+++ src/main/java/ru/homework/hometask07/dao/RoleRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import ru.homework.hometask07.dao.entity.RoleEntity;
 @Repository
-public interface RoleRepository extends JpaRepository<RoleEntity, Integer> {
+public interface RoleRepository extends JpaRepository<RoleEntity, Long> {
 }
